requirements.txt
fastapi>=0.110
uvicorn[standard]>=0.27
pydantic>=2.6
pydantic-settings>=2.2
python-dotenv>=1.0
SQLAlchemy>=2.0
requests>=2.31
passlib>=1.7.4
python-jose[cryptography]>=3.3.0
cryptography>=42.0
boto3>=1.34
